requirements.txt
asgiref==3.4.1
Django==4.0b1
django-cors-headers==3.10.0
djangorestframework==3.12.4
pkg_resources==0.0.0
psycopg2-binary==2.9.2
python-dotenv==0.17.1
pytz==2021.3
sqlparse==0.4.2